calorie_tracker/backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
httpx==0.24.0
python-multipart==0.0.6
firebase-admin==6.1.0
openai==0.28.0
python-dotenv==1.0.0
